Rename player-proxy-machines to wip
